...G...... [DEV-802] type cast to 'int'

+++ src/libs/zbxsysinfo/common/http.c
@@ -42,14 +42,14 @@ static int	get_http_page(const char *host, const char *path, unsigned short port
 	if (NULL != (wrong_chr = strpbrk(host, URI_PROHIBIT_CHARS)))
 	{
 		*error = zbx_dsprintf(NULL, "Invalid characters are present after first part of hostname:%.*s",
-				wrong_chr - path, host);
+				(int)(wrong_chr - path), host);
 		return SYSINFO_RET_FAIL;
 	}
 
 	if (NULL != (wrong_chr = strpbrk(path, URI_PROHIBIT_CHARS)))
 	{
 		*error = zbx_dsprintf(NULL, "Invalid characters are present after first part of path:%.*s",
-				wrong_chr - path, path);
+				(int)(wrong_chr - path), path);
 		return SYSINFO_RET_FAIL;
 	}
